+++ src/append.js
@@ -1,8 +1,22 @@
-module.exports = function append (tree, node) {
+const template = require('./template')
+
+function getNode (input) {
+  if (typeof input === 'string') return template(input)
+  return input
+}
+
+function addNode (tree, node) {
+  if (Array.isArray(node)) return tree.concat(node)
+  tree.push(node)
+  return tree
+}
+
+module.exports = function append (tree, input) {
+  const node = getNode(input)
   if (Array.isArray(tree)) {
-    tree.push(node)
+    tree = addNode(tree, node)
   } else if (Array.isArray(tree.body)) {
-    tree.body.push(node)
+    tree.body = addNode(tree.body, node)
   }
   return tree
 }